+++ packages/orq-rc/RELEASES.md
@@ -1099,3 +1099,13 @@ Based on:
 - [typescript v3.3.0-rc.12] packages/orq-rc
 ### Releases
 - [NPM v3.3.0-rc.12] https://www.npmjs.com/package/@orq-ai/node/v/3.3.0-rc.12 - packages/orq-rc
+
+## 2025-03-27 22:16:15
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.525.1 (2.562.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.3.0-rc.13] packages/orq-rc
+### Releases
+- [NPM v3.3.0-rc.13] https://www.npmjs.com/package/@orq-ai/node/v/3.3.0-rc.13 - packages/orq-rc